modal-all-works
Check this, in my opinion it's okey
