+++ src/css/main.css
@@ -1588,12 +1588,14 @@ img {
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
   -webkit-box-align: center;
       -ms-flex-align: center;
           align-items: center;
+  position: relative;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
 }
 
 .modal {
@@ -1616,57 +1618,50 @@ img {
 .modal .wrapper .close {
   position: absolute;
   right: 25px;
-  top: 25px;
+  top: 5px;
   width: 17px;
   height: 17px;
 }
 
-.modal .wrapper .imgRound {
-  opacity: 0;
-  overflow: hidden;
-  width: 102px;
-  height: 102px;
-  margin: 75px auto 18px;
-  background-image: url("./../images/temporary/pic.png");
-  background-position: 50% 50%;
-  border-radius: 50% 50%;
+.modal .wrapper .imgRound img {
+  width: 100px;
+  height: 100px;
 }
 
 @media screen and (min-width: 768px) {
-  .modal .wrapper .imgRound {
-    width: 70px;
-    height: 70px;
-    margin: 28px auto 43px;
+  .modal .wrapper .imgRound img {
+    width: 110px;
+    height: 110px;
   }
 }
 
 @media screen and (min-width: 1280px) {
-  .modal .wrapper .imgRound {
+  .modal .wrapper .imgRound img {
     width: 132px;
     height: 132px;
-    margin: 62px auto 41px;
   }
 }
 
 .modal .wrapper {
   width: 272px;
-  height: 915px;
-  border-radius: 150px 0px;
+  height: 100vh;
+  border-radius: 150px 0px 150px 0px;
   background-color: #fff2fd;
 }
 
 @media screen and (min-width: 768px) {
   .modal .wrapper {
     width: 595px;
-    height: 831px;
+    height: 100vh;
     border-radius: 300px 0px;
+    padding-top: 30px;
   }
 }
 
 @media screen and (min-width: 1280px) {
   .modal .wrapper {
     width: 930px;
-    height: 1023px;
+    height: 100vh;
   }
 }
 
@@ -1721,15 +1716,24 @@ img {
     display: block;
     width: 233px;
     height: 416px;
-    margin: 0 30px 0 50px;
+    padding-right: 20px;
+    padding-top: 50px;
   }
 }
 
 @media screen and (min-width: 1280px) {
   .modal-list__img {
-    width: 360px;
-    height: 641px;
+    width: 300px;
+    height: 500px;
     margin: 0 60px 0 30px;
+    -webkit-box-align: center;
+        -ms-flex-align: center;
+            align-items: center;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    padding-right: 0;
+    padding-top: 0;
   }
 }
 
@@ -1752,14 +1756,17 @@ img {
   font-weight: 400;
   font-size: 14px;
   line-height: 1.25;
-  margin: 5px 0 10px 0;
+  margin: 8px 0;
+  max-height: 82px;
+  overflow: hidden;
 }
 
 @media screen and (min-width: 1280px) {
   .modal-list__text {
-    font-size: 18px;
+    font-size: 14px;
     line-height: 1.22;
-    margin: 7px 0 20px 0;
+    max-height: 70px;
+    overflow: hidden;
   }
 }
 
@@ -1774,7 +1781,7 @@ img {
 }
 
 .modal-list__text--label::before {
-  content: '';
+  content: "";
   display: inline-block;
   background: url("./../images/modal/ticket.png");
   width: 24px;
@@ -1797,7 +1804,6 @@ img {
   cursor: pointer;
   width: 145px;
   height: 40px;
-  margin-bottom: 20px;
   border-radius: 5px;
   border-style: none;
 }
@@ -1910,4 +1916,3 @@ img {
 .over {
   overflow: hidden;
 }
-/*# sourceMappingURL=main.css.map */
