+++ src/sass/layout/_modal.scss
@@ -1,8 +1,9 @@
 @import "../utils/variable";
-.div-modal{
+.div-modal {
   display: flex;
-  justify-content: center;
   align-items: center;
+  position: relative;
+  flex-direction: column;
 }
 .modal {
   display: flex;
@@ -15,51 +16,43 @@
     .close {
       position: absolute;
       right: 25px;
-      top: 25px;
+      top: 5px;
       width: 17px;
       height: 17px;
     }
 
-    .imgRound {
-      opacity:0;
-      overflow: hidden;
-      width: 102px;
-      height: 102px;
-      margin: 75px auto 18px;
-
-      background-image: url("./../images/temporary/pic.png");
-      background-position: 50% 50%;
-      border-radius: 50% 50%;
+    .imgRound img {
+      width: 100px;
+      height: 100px;
 
       @media screen and (min-width: 768px) {
-        width: 70px;
-        height: 70px;
-        margin: 28px auto 43px;
+        width: 110px;
+        height: 110px;
       }
 
       @media screen and (min-width: 1280px) {
         width: 132px;
         height: 132px;
-        margin: 62px auto 41px;
       }
     }
   }
 
   .wrapper {
     width: 272px;
-    height: 915px;
-    border-radius: 150px 0px;
+    height: 100vh;
+    border-radius: 150px 0px 150px 0px;
     background-color: getColor(bg-modal);
 
     @media screen and (min-width: 768px) {
       width: 595px;
-      height: 831px;
+      height: 100vh;
       border-radius: 300px 0px;
+      padding-top: 30px;
     }
 
     @media screen and (min-width: 1280px) {
       width: 930px;
-      height: 1023px;
+      height: 100vh;
     }
   }
 
@@ -85,20 +78,24 @@
     &__img {
       visibility: hidden;
       display: none;
-      
 
       @media screen and (min-width: 768px) {
         visibility: visible;
         display: block;
         width: 233px;
         height: 416px;
-        margin: 0 30px 0 50px;
+        padding-right: 20px;
+        padding-top: 50px;
       }
 
       @media screen and (min-width: 1280px) {
-        width: 360px;
-        height: 641px;
+        width: 300px;
+        height: 500px;
         margin: 0 60px 0 30px;
+        align-items: center;
+        display: flex;
+        padding-right: 0;
+        padding-top: 0;
       }
     }
 
@@ -118,12 +115,15 @@
       font-weight: 400;
       font-size: 14px;
       line-height: 1.25;
-      margin: 5px 0 10px 0;
+      margin: 8px 0;
+      max-height: 82px;
+      overflow: hidden;
 
       @media screen and (min-width: 1280px) {
-        font-size: 18px;
+        font-size: 14px;
         line-height: 1.22;
-        margin: 7px 0 20px 0;
+        max-height: 70px;
+        overflow: hidden;
       }
 
       &-width {
@@ -134,9 +134,9 @@
       }
 
       &--label::before {
-        content: '';
+        content: "";
         display: inline-block;
-        background: url('./../images/modal/ticket.png');
+        background: url("./../images/modal/ticket.png");
         // fill:red;
         width: 24px;
         height: 16px;
@@ -158,7 +158,7 @@
         cursor: pointer;
         width: 145px;
         height: 40px;
-        margin-bottom: 20px;
+
         border-radius: 5px;
         border-style: none;
 
@@ -198,7 +198,6 @@
         line-height: 1.25;
 
         @media screen and (min-width: 768px) {
-          
           bottom: 56px;
           left: 170px;
         }
@@ -217,39 +216,39 @@
   }
 }
 
-.backdrop{
-    visibility: visible;
-    background: rgba(32, 13, 29, 0.6);
-backdrop-filter: blur(3px);
-    position: fixed;
-    top: 0;
-    left: 0;
-    opacity:1;
-    width: 100%;
-    height: 100%;
-    transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1),filter 250ms cubic-bezier(0.4, 0, 0.2, 1);
- 
-    overflow: hidden;
+.backdrop {
+  visibility: visible;
+  background: rgba(32, 13, 29, 0.6);
+  backdrop-filter: blur(3px);
+  position: fixed;
+  top: 0;
+  left: 0;
+  opacity: 1;
+  width: 100%;
+  height: 100%;
+  transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1),
+    filter 250ms cubic-bezier(0.4, 0, 0.2, 1);
+
+  overflow: hidden;
 }
-.is-hidden{
-   pointer-events: none;
-    opacity:0; 
+.is-hidden {
+  pointer-events: none;
+  opacity: 0;
 }
-.wrapper{
-   
-    filter: blur(0);
-    position: fixed;
-    top:0;
-    left:50%;
-    transform: translateX(-50%);
-    width:80%;
-    height:80vh;
-  
-    border:3px solid white;
+.wrapper {
+  filter: blur(0);
+  position: fixed;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 80%;
+  height: 80vh;
+
+  border: 3px solid white;
 }
-.wrapper-active{
-    display:block;
+.wrapper-active {
+  display: block;
 }
-.over{
-    overflow: hidden;
+.over {
+  overflow: hidden;
 }
